boards: 54h20dk: Changed to run with BT stack from local RAM
Adjust the ISR vectors table offset to make sure mcuboot
accesser right location. The vectors table is implicitly
adjusted to be aligned on 0x800 dividable address.
mcuboot must be informed about the adjustment.

NOTE: This is 54h20 use case specific change matching the
address of the code in local RAM.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>